owsnompy: Clean up temps and prints
